fixed IRC20 func, change ERC20 contact cleanup, deleted unsed data, change custom token contact add to constructon more data to create contract

+++ contracts/IERC20.sol
@@ -12,7 +12,7 @@ interface IERC20 {
 
   function balanceOf(address account) external view returns (uint256);
 
-  function transfer(address recipient, uint64 amount) external returns (bool);
+  function transfer(address recipient, uint256 amount) external returns (bool);
 
   function allowance(address owner, address spender) external view returns (uint256);
